changed return type of api's from uint8_t to uint32_t

+++ src/api/heap.h
@@ -11,7 +11,7 @@ extern "C" {
 
 extern void heap_init();
 extern void* heap_allocate(uint32_t size);
-extern uint8_t heap_release(void *ptr);
+extern uint32_t heap_release(void *ptr);
 
 #ifdef	__cplusplus
 }
