+++ src/api/umpu.h
@@ -9,8 +9,8 @@
 extern "C" {
 #endif
 
-extern uint8_t user_mpu_enable(uint8_t region, uint32_t address, uint32_t attributes);
-extern uint8_t user_mpu_disable(uint8_t region);
+extern uint32_t user_mpu_enable(uint8_t region, uint32_t address, uint32_t attributes);
+extern uint32_t user_mpu_disable(uint8_t region);
 
 #ifdef __cplusplus
 }
