+++ src/share.c
@@ -17,7 +17,7 @@
 extern uint32_t arch_semaphore_inc(uint32_t *semaphore);
 extern uint32_t arch_semaphore_dec(uint32_t *semaphore);
 
-uint8_t spin_lock(uint32_t *mutex_st, uint16_t retry){
+uint32_t spin_lock(uint32_t *mutex_st, uint16_t retry){
 
     uint8_t state;
     uint16_t loop=0;
@@ -81,7 +81,7 @@ uint8_t spin_lock(uint32_t *mutex_st, uint16_t retry){
 
 }
 
-uint8_t semaphore_increment(uint32_t *semaphore, uint16_t retry) {
+uint32_t semaphore_increment(uint32_t *semaphore, uint16_t retry) {
 
     uint32_t state;
     uint16_t loop=0;
@@ -107,7 +107,7 @@ uint8_t semaphore_increment(uint32_t *semaphore, uint16_t retry) {
     return ERROR_NONE;
 }
 
-uint8_t semaphore_decrement(uint32_t *semaphore, uint16_t retry) {
+uint32_t semaphore_decrement(uint32_t *semaphore, uint16_t retry) {
 
     uint8_t state;
     uint16_t loop=0;
